requirements.txt
django~=4.0.6
Pillow~=9.2.0
djangorestframework~=3.14.0